Allow deletion and renaming on exFAT
Essentially there was a bug when editing swp files on my external hard drive, which is formatted exFAT. The unlink leaves a "vestige" of the file if it is still open with a valid file descriptor, where the "ls" command still sees that the file is there but knows nothing about it. Because of this, the subsequent rename function call fails. This did not occur on NTFS or ext4. Basically the only change here is making sure the file is closed prior to deletion, so that there is no remaining "vestige", and the rename does not end in an error.